Update to handle command line options properly.
contestest.py has to be in the top level to capture custom command
line flags, so we had to move that.

Tests were asking for symbolic refs, which was fragile: sometimes
we'd get valid symbolic ref which was not what we were looking for.
Ensuring we checked out the proper hash makes that more robust.