Add test of retrieving a value using a namespace
This is the first test which has discovered a bug in our program!
Apparently the CLI library which the chaos tool uses can't deal
with values which contain an '=' char. This makes things tough when
one of your values is expected to be base64-encoded data.

There's a workaround, but until chaostool's branch
prgn-updates-for-integration-tests is merged, that branch is required
for these tests to pass.