Finish current work on integration tests.
This adds a few tests, and skips others. The tests with unconditional
skips have them because they are incomplete, and sometimes fail for
reasons having more to do with their implementation than the code
under test.

With more time, we'd likely be able to get these all done and complete
the planned program of testing, but that time isn't available: we're
now starting a new development sprint. We have to draw a line under
it somewhere, so that line is here.