change version check to just verify connection
it used to do `ndau("-v version check")`, which verified that the ndau version on the server was identical to that on the client. This policy was overly restrictive: we're making plenty of non-breaking changes these days. Better to just do `version remote`, which has all of the benefits as far as testing the data flow without causing spurious failures if the ndau tool is at a more recent version than testnet.