Make SIB and CVC integration tests work
The SIB test was just a matter of not attempting to test the behavior
of the SIB script, instead just trusting its unit tests to check its
math.

The CVC test fix was substantially more complicated, because we've made it
substantially harder to test CVC. The solution here works on localnet,
and will be skipped on other nets; that's because other nets just don't
have enough information and context to properly construct a CVC tx which
could be valid.

If the test is skipped and it's not clear why, run pytest with the
`-rs` flag to display a short summary of reasons why tests were skipped.